media: attach network storage
To keep the username and password secret, created a secrets.nix file
that will be stored on /etc/nixos/* on the machine. This file, for
obvious reasons, will be created on the machine and ignored by git so
that these creds stay private. Will find a better way to do this later.